Enable language detection and URL modifications
Hi @soderlind, 
I found a small modification that could be done to enable these two useful Polylang options (language detection and URL modifications). 
I think "force_lang" and "language detection" in the homepage should be disabled only when we are in the customizer otherwise they doesn't work in the frontend! 

thanks for your plugin, it was very helpful!

Erik